Move vertex definitions.
