--- demo/src/features/tile_layer/tile_layer_vertex.rs
@@ -1,7 +1,0 @@
-/// Vertex format for vertices sent to the GPU
-#[derive(Clone, Debug, Copy, Default)]
-#[repr(C)]
-pub struct TileLayerVertex {
-    pub position: [f32; 3],
-    pub uv: [f32; 2],
-}
